+++ src/stores/AppStore/AppStore.ts
@@ -1659,6 +1659,15 @@ export class AppStore {
         }
         this.handleThemeChange(mediaQuery.matches);
 
+        // Add spellcheck="false" in input elements with no spellcheck attribute
+        document.addEventListener("input", (ev: Event) => {
+            const target = ev.target;
+            const isInput = target && target instanceof HTMLInputElement;
+            if (isInput && !target.getAttribute("spellcheck")) {
+                target.setAttribute("spellcheck", "false");
+            }
+        });
+
         // Display toasts when connection status changes
         autorun(() => {
             const newConnectionStatus = this.backendService.connectionStatus;